chore(dashboards): remove feature flag paginatedDashboards
