+++ src/dashboards/components/dashboard_index/DashboardCards.tsx
@@ -1,22 +1,19 @@
 // Libraries
 import React, {PureComponent} from 'react'
 import {connect, ConnectedProps} from 'react-redux'
-import memoizeOne from 'memoize-one'
 
 // Components
 import DashboardCard from 'src/dashboards/components/dashboard_index/DashboardCard'
-import {ResourceCard, TechnoSpinner} from '@influxdata/clockface'
 import AssetLimitAlert from 'src/cloud/components/AssetLimitAlert'
-
-// Selectors
-import {getSortedResources, SortTypes} from 'src/shared/utils/sort'
-import {getMe} from 'src/me/selectors'
-import {getOrg} from 'src/organizations/selectors'
+import {ResourceCard} from '@influxdata/clockface'
 
 // Types
-import {AppState, Dashboard, RemoteDataState} from 'src/types'
-import {Sort} from 'src/clockface'
+import {AppState, Dashboard} from 'src/types'
 import {LimitStatus} from 'src/cloud/actions/limits'
+
+// Selectors
+import {getMe} from 'src/me/selectors'
+import {getOrg} from 'src/organizations/selectors'
 
 // Contexts
 import {
@@ -47,9 +44,6 @@ interface StateProps {
 
 interface OwnProps {
   dashboards: Dashboard[]
-  sortKey: string
-  sortDirection: Sort
-  sortType: SortTypes
   onFilterChange: (searchTerm: string) => void
 }
 
@@ -57,20 +51,12 @@ type ReduxProps = ConnectedProps<typeof connector>
 type Props = OwnProps & StateProps & ReduxProps
 
 class DashboardCards extends PureComponent<Props> {
-  private _observer
   private _isMounted = true
-  private _spinner
   private _assetLimitAlertStyle = {
     height: 'inherit',
   }
 
-  private memGetSortedResources = memoizeOne<typeof getSortedResources>(
-    getSortedResources
-  )
-
   state = {
-    pages: 1,
-    windowSize: 0,
     pinnedItems: [],
     dashboardCardHeight: 'inherit',
   }
@@ -79,7 +65,6 @@ class DashboardCards extends PureComponent<Props> {
     if (isFlagEnabled('pinnedItems') && CLOUD) {
       this.updatePinnedItems()
     }
-    this.setState(prev => ({...prev, windowSize: 15}))
   }
 
   public componentDidUpdate() {
@@ -97,41 +82,41 @@ class DashboardCards extends PureComponent<Props> {
     this._isMounted = false
   }
 
-  private registerSpinner = elem => {
-    this._spinner = elem
+  public render() {
+    const {dashboards, onFilterChange} = this.props
 
-    if (!elem) {
-      return
-    }
+    const {pinnedItems} = this.state
 
-    let count = 1.0
-    const threshold = []
-
-    while (count > 0) {
-      threshold.push(count)
-      count -= 0.1
-    }
-
-    threshold.reverse()
-
-    this._observer = new IntersectionObserver(this.measure, {
-      threshold,
-      rootMargin: '60px 0px',
-    })
-
-    this._observer.observe(this._spinner)
-  }
-
-  private measure = entries => {
-    if (
-      entries
-        .map(e => e.isIntersecting)
-        .reduce((prev, curr) => prev || curr, false)
-    ) {
-      this.setState({
-        pages: this.state.pages + 1,
-      })
-    }
+    return (
+      <div className="dashboards-card-grid">
+        {dashboards.map(({id, name, description, labels, meta}) => (
+          <DashboardCard
+            key={id}
+            id={id}
+            name={name}
+            labels={labels}
+            updatedAt={meta.updatedAt}
+            description={description}
+            onFilterChange={onFilterChange}
+            onPinDashboard={this.handlePinDashboard}
+            onUnpinDashboard={this.handleUnpinDashboard}
+            isPinned={
+              !!pinnedItems.find(item => item?.metadata.dashboardID === id)
+            }
+          />
+        ))}
+        {this.props.limitStatus === 'exceeded' && (
+          <ResourceCard style={{height: this.state.dashboardCardHeight}}>
+            <AssetLimitAlert
+              className="dashboards--asset-alert"
+              resourceName="dashboards"
+              limitStatus={this.props.limitStatus}
+              style={this._assetLimitAlertStyle}
+            />
+          </ResourceCard>
+        )}
+      </div>
+    )
   }
 
   public updatePinnedItems = () => {
@@ -179,68 +164,6 @@ class DashboardCards extends PureComponent<Props> {
     } catch (err) {
       this.props.sendNotification(pinnedItemFailure(err.message, 'delete'))
     }
-  }
-
-  public render() {
-    const {
-      dashboards,
-      sortDirection,
-      sortKey,
-      sortType,
-      onFilterChange,
-    } = this.props
-    const sortedDashboards = this.memGetSortedResources(
-      dashboards,
-      sortKey,
-      sortDirection,
-      sortType
-    )
-
-    const {windowSize, pages, pinnedItems} = this.state
-
-    return (
-      <div style={{height: '100%'}}>
-        <div className="dashboards-card-grid">
-          {sortedDashboards
-            .filter(d => d.status === RemoteDataState.Done)
-            .slice(0, pages * windowSize)
-            .map(({id, name, description, labels, meta}) => (
-              <DashboardCard
-                key={id}
-                id={id}
-                name={name}
-                labels={labels}
-                updatedAt={meta.updatedAt}
-                description={description}
-                onFilterChange={onFilterChange}
-                onPinDashboard={this.handlePinDashboard}
-                onUnpinDashboard={this.handleUnpinDashboard}
-                isPinned={
-                  !!pinnedItems.find(item => item?.metadata.dashboardID === id)
-                }
-              />
-            ))}
-          {this.props.limitStatus === 'exceeded' && (
-            <ResourceCard style={{height: this.state.dashboardCardHeight}}>
-              <AssetLimitAlert
-                className="dashboards--asset-alert"
-                resourceName="dashboards"
-                limitStatus={this.props.limitStatus}
-                style={this._assetLimitAlertStyle}
-              />
-            </ResourceCard>
-          )}
-        </div>
-        {windowSize * pages < dashboards.length && (
-          <div
-            style={{height: '140px', margin: '14px auto'}}
-            ref={this.registerSpinner}
-          >
-            <TechnoSpinner />
-          </div>
-        )}
-      </div>
-    )
   }
 }
